W-11742823: Review registry objects needed only by “fat” tooling
* Split tooling test from AbstractDbIntegrationTestCase
