+++ src/test/java/org/mule/extension/db/integration/delete/DeleteMetadataTestCase.java
@@ -10,8 +10,9 @@ package org.mule.extension.db.integration.delete;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-import org.mule.extension.db.integration.AbstractDbIntegrationTestCase;
+import static org.hamcrest.MatcherAssert.assertThat;
+
+import org.mule.extension.db.integration.AbstractDbMetadataIntegrationTestCase;
 import org.mule.metadata.api.model.ArrayType;
 import org.mule.metadata.api.model.MetadataType;
 import org.mule.metadata.api.model.NullType;
@@ -22,7 +23,7 @@ import org.mule.runtime.api.metadata.resolving.MetadataResult;
 
 import org.junit.Test;
 
-public class DeleteMetadataTestCase extends AbstractDbIntegrationTestCase {
+public class DeleteMetadataTestCase extends AbstractDbMetadataIntegrationTestCase {
 
   @Override
   protected String[] getFlowConfigurationResources() {
