+++ src/test/java/org/mule/extension/db/integration/select/SelectMetadataInputTestCase.java
@@ -13,15 +13,16 @@
 
 package org.mule.extension.db.integration.select;
 
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.instanceOf;
-import static org.hamcrest.Matchers.not;
-import static org.hamcrest.core.Is.is;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assume.assumeThat;
 import static org.mule.extension.db.integration.DbTestUtil.DbType.MYSQL;
 
-import org.mule.extension.db.integration.AbstractDbIntegrationTestCase;
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.instanceOf;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.not;
+import static org.hamcrest.core.Is.is;
+import static org.junit.Assume.assumeThat;
+
+import org.mule.extension.db.integration.AbstractDbMetadataIntegrationTestCase;
 import org.mule.metadata.api.model.NullType;
 import org.mule.metadata.api.model.ObjectFieldType;
 import org.mule.metadata.api.model.ObjectType;
@@ -33,7 +34,7 @@ import java.util.Optional;
 
 import org.junit.Test;
 
-public class SelectMetadataInputTestCase extends AbstractDbIntegrationTestCase {
+public class SelectMetadataInputTestCase extends AbstractDbMetadataIntegrationTestCase {
 
   @Override
   protected String[] getFlowConfigurationResources() {
